Support Apple silicon on MacBook

+++ Container-Root/eks/ops/setup/install-aws-cli.sh
@@ -7,14 +7,23 @@ ARCH=$(uname -m)
 OS=$(uname -s)
 OS_LOWER=$(echo $OS | tr '[:upper:]' '[:lower:]')
 PLATFORM=${OS_LOWER}-${ARCH}
-URL=https://awscli.amazonaws.com/awscli-exe-${PLATFORM}.zip
-echo "$URL"
-
 # Install aws cli
-curl "$URL" -o "awscliv2.zip"
-unzip awscliv2.zip
-./aws/install
-sudo cp -f /usr/local/bin/aws /bin/aws
+if [ "$OS" == "Darwin" ]; then
+	echo "Installing AWS CLI on Mac ..."
+	URL="https://awscli.amazonaws.com/AWSCLIV2.pkg"
+	echo "${URL}"
+	curl "${URL}" -o "AWSCLIV2.pkg"
+	sudo installer -pkg AWSCLIV2.pkg -target /
+else
+	echo "Installing AWS CLI on Linux ..."
+	URL=https://awscli.amazonaws.com/awscli-exe-${PLATFORM}.zip
+	echo "$URL"
+	curl "$URL" -o "awscliv2.zip"
+	unzip awscliv2.zip
+	./aws/install
+	sudo cp -f /usr/local/bin/aws /bin/aws
+	rm -rf ./aws
+	rm -f awscliv2.zip
+fi
 aws --version
-rm -rf ./aws
-rm -f awscliv2.zip
+
